+++ src/main/java/rbasamoyai/createbigcannons/cannon_control/carriage/CannonCarriageEntity.java
@@ -372,11 +372,6 @@ public class CannonCarriageEntity extends Entity implements ControlPitchContrapt
         }
     }
 
-//    @Override
-//	public double getPassengersRidingOffset() {
-//		return 27 / 32f;
-//	} fixme above, don't know how this works exactly --ritchie
-
 	@Override
 	public boolean canBeCollidedWith() {
 		return true;
@@ -425,7 +420,16 @@ public class CannonCarriageEntity extends Entity implements ControlPitchContrapt
 			this.xRotO = flag ? this.cannonContraption.prevPitch : -this.cannonContraption.prevPitch;
 			this.setXRot(flag ? this.cannonContraption.pitch : -this.cannonContraption.pitch);
 			this.yRotO = this.yRotO + Mth.degreesDifference(this.yRotO, this.cannonContraption.prevYaw);
-			this.setYRot(this.getYRot() + Mth.degreesDifference(this.getYRot(), this.cannonContraption.yaw));
+            float deltaYaw = Mth.degreesDifference(this.getYRot(), this.cannonContraption.yaw);
+			this.setYRot(this.getYRot() + deltaYaw);
+
+            if (this.level().isClientSide) {
+                Vector4f newState = this.getWheelState();
+                float f = deltaYaw * 7;
+                newState.add(f, f, f, f);
+                this.setWheelState(newState);
+                NetworkPlatform.sendToServer(ServerboundCarriageWheelPacket.entity(this));
+            }
 		}
 	}